admin_microservice/pyproject.toml
[project]
name = "admin-microservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcus",email = "marc303i@stud.kea.dk"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pika (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
]
packages = [
    { include = "admin-microservice" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
